pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "honeybee-ml"
version = "0.1.0"
description = "A Scalable Modular Framework for Multimodal AI in Oncology"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Aakash Tripathi", email = "aakash.tripathi@moffitt.org"},
    {name = "Lab Rasool", email = "ghulam.rasool@moffitt.org"}
]
maintainers = [
    {name = "Aakash Tripathi", email = "aakash.tripathi@moffitt.org"}
]
keywords = [
    "multimodal AI",
    "oncology",
    "cancer research",
    "medical imaging",
    "clinical NLP",
    "machine learning",
    "pathology",
    "biomedical",
    "healthcare"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Text Processing :: Linguistic",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.40.0",
    "scikit-image>=0.19.0",
    "scipy>=1.7.0",
    "matplotlib>=3.3.0",
    "imageio>=2.9.0",
    "tqdm>=4.60.0",
    "Pillow>=9.0.0",
    "opencv-python>=4.5.0",
    "huggingface_hub>=0.16.0",
]

[project.optional-dependencies]
# Clinical text processing dependencies
clinical = [
    # Document ingestion
    "unstructured[pdf,image]>=0.10.0",
    "pymupdf>=1.23.0",
    "langchain>=0.1.0",
    "pdf2image>=1.16.0",
    "PyPDF2>=3.0.0",
    # Clinical NLP
    "scispacy>=0.5.0",
    "spacy>=3.5.0",
    # Interop
    "fhir.resources>=7.0.0",
    "hl7apy>=1.3.0",
    # Ontology APIs
    "requests>=2.28.0",
    # Temporal
    "dateparser>=1.1.0",
    # Embeddings API
    "litellm>=1.0.0",
    # OCR (system dep: tesseract-ocr)
    "pytesseract>=0.3.8",
]

# Heavy optional extras for clinical
clinical-full = [
    "honeybee-ml[clinical]",
    "medspacy>=1.0.0",
    "medcat>=1.9.0",
]

# Pathology processing dependencies
pathology = [
    "openslide-python>=1.2.0",
    "colour-science>=0.4.0",
    "albumentations>=1.3.0",
    "cucim>=23.0.0",
    "umap-learn>=0.5.0",
    "scikit-learn>=1.0.0",
]

# Radiology image processing
radiology = [
    "SimpleITK>=2.2.0",
    "monai>=1.0.0",
    "torchio>=0.19.0",
    "pydicom>=2.3.0",
    "nibabel>=5.0.0",
    "lungmask>=0.2.0",
    "TotalSegmentator>=2.0.0",
    "dipy>=1.7.0",
    "intensity-normalization>=2.0.0",
    "neuroCombat>=0.2.0",
    "nnunetv2>=2.0",
]

# Molecular data processing
molecular = [
    "pyarrow>=10.0.0",
    "fastparquet>=2023.0.0",
]

# Database connectivity
database = [
    "pymongo>=4.0.0",
]

# Development and testing
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# Visualization and interactive tools
visualization = [
    "ipykernel>=6.0.0",
    "ipywidgets>=8.0.0",
]

# AI/ML model extras
models = [
    "timm>=0.9.0",
    "onnxruntime>=1.15.0",
    "peft>=0.5.0",
]

# Complete installation with all features
all = [
    "honeybee-ml[clinical-full,pathology,radiology,molecular,database,visualization,models]",
]

[project.urls]
Homepage = "https://github.com/lab-rasool/HoneyBee"
Documentation = "https://lab-rasool.github.io/HoneyBee/"
Repository = "https://github.com/lab-rasool/HoneyBee"
"Bug Tracker" = "https://github.com/lab-rasool/HoneyBee/issues"
"Paper" = "https://arxiv.org/abs/2405.07460"

[tool.setuptools.packages.find]
where = ["."]
include = ["honeybee*"]
exclude = ["tests*", "docs*", "website*", "results*", "BACKUP*", "to_delete*", "paper*"]

[tool.setuptools.package-data]
honeybee = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.tox
  | build
  | dist
  | __pycache__
  | \.eggs
  | \.pytest_cache
)/
'''

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501", "N813"]  # E501: Line too long (handled by black), N813: SimpleITK as sitk

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov=honeybee --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
